requirements.txt
pandas
openai
crhomadb
python-dotenv
tqdm